Fix rebase issues.
