+++ src/main/kotlin/util/iroha/Hasher.kt
@@ -10,21 +10,21 @@ import java.math.BigInteger
  */
 fun getHash(data: ByteArray): String {
     val tx = iroha.protocol.BlockOuterClass.Transaction.parseFrom(data)
-    val creator = tx.payload.creatorAccountId
-    val quorum = tx.payload.quorum
-    val time = tx.payload.createdTime
-    if (tx.payload.commandsList.size != 1) {
+    val creator = tx.payload.reducedPayload.creatorAccountId
+    val quorum = tx.payload.reducedPayload.quorum
+    val time = tx.payload.reducedPayload.createdTime
+    if (tx.payload.reducedPayload.commandsList.size != 1) {
         return ""
     }
-    if (!tx.payload.commandsList[0].hasTransferAsset()) {
+    if (!tx.payload.reducedPayload.commandsList[0].hasTransferAsset()) {
         return ""
     }
-    val cmd = tx.payload.commandsList[0].transferAsset
+    val cmd = tx.payload.reducedPayload.commandsList[0].transferAsset
 
     val hash = ModelTransactionBuilder()
             .creatorAccountId(creator)
             .createdTime(BigInteger.valueOf(time))
-            .quorum(quorum.toLong())
+            .quorum(quorum)
             .transferAsset(
                     cmd.srcAccountId,
                     cmd.destAccountId,
