+++ src/main/resources/defaults_deploy.properties
@@ -43,7 +43,7 @@ relayRegistration.ethGasPrice=1
 # Gas limit for relay contract deployment
 relayRegistration.ethGasLimit=999999
 # address of master contract
-ethereum.masterAddress=0xb1c73070b129a6ce1a2e028e5a62fca2f48e8866
+ethereum.masterAddress=0x7b301aa25285ecfb00ac160484415b9ee5386d6e
 # ========= Registration Service =========
 # port of registration service
 registrationService.port=8083
